+++ govcd/orgvdcnetwork_test.go
@@ -9,8 +9,9 @@ package govcd
 import (
 	"fmt"
 
-	"github.com/vmware/go-vcloud-director/v2/types/v56"
 	. "gopkg.in/check.v1"
+
+	"github.com/vmware/go-vcloud-director/v2/types/v56"
 )
 
 func (vcd *TestVCD) Test_NetRefresh(check *C) {
@@ -39,20 +40,20 @@ func (vcd *TestVCD) Test_NetRefresh(check *C) {
 
 }
 
-func (vcd *TestVCD) Test_CreateOrgVdcNetworkRouted(check *C) {
-	vcd.testCreateOrgVdcNetworkRouted(check, "10.10.101", false, false)
+func (vcd *TestVCD) Test_CreateUpdateOrgVdcNetworkRouted(check *C) {
+	vcd.testCreateUpdateOrgVdcNetworkRouted(check, "10.10.101", false, false)
 }
 
-func (vcd *TestVCD) Test_CreateOrgVdcNetworkRoutedSubInterface(check *C) {
-	vcd.testCreateOrgVdcNetworkRouted(check, "10.10.102", true, false)
+func (vcd *TestVCD) Test_CreateUpdateOrgVdcNetworkRoutedSubInterface(check *C) {
+	vcd.testCreateUpdateOrgVdcNetworkRouted(check, "10.10.102", true, false)
 }
 
-func (vcd *TestVCD) Test_CreateOrgVdcNetworkRoutedDistributed(check *C) {
-	vcd.testCreateOrgVdcNetworkRouted(check, "10.10.103", false, true)
+func (vcd *TestVCD) Test_CreateUpdateOrgVdcNetworkRoutedDistributed(check *C) {
+	vcd.testCreateUpdateOrgVdcNetworkRouted(check, "10.10.103", false, true)
 }
 
-// Tests the creation of an Org VDC network connected to an Edge Gateway
-func (vcd *TestVCD) testCreateOrgVdcNetworkRouted(check *C, ipSubnet string, subInterface, distributed bool) {
+// Tests the creation and update of an Org VDC network connected to an Edge Gateway
+func (vcd *TestVCD) testCreateUpdateOrgVdcNetworkRouted(check *C, ipSubnet string, subInterface, distributed bool) {
 	fmt.Printf("Running: %s\n", check.TestName())
 	networkName := TestCreateOrgVdcNetworkRouted
 
@@ -140,6 +141,13 @@ func (vcd *TestVCD) testCreateOrgVdcNetworkRouted(check *C, ipSubnet string, sub
 	check.Assert(network, NotNil)
 	check.Assert(network.OrgVDCNetwork.Name, Equals, networkName)
 	check.Assert(network.OrgVDCNetwork.Description, Equals, networkDescription)
+	check.Assert(network.OrgVDCNetwork.Configuration, NotNil)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes, NotNil)
+	check.Assert(len(network.OrgVDCNetwork.Configuration.IPScopes.IPScope), Not(Equals), 0)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].Gateway, Equals, gateway)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].StartAddress, Equals, startAddress)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].EndAddress, Equals, endAddress)
+
 	if subInterface {
 		check.Assert(network.OrgVDCNetwork.Configuration.SubInterface, NotNil)
 		check.Assert(*network.OrgVDCNetwork.Configuration.SubInterface, Equals, true)
@@ -150,6 +158,34 @@ func (vcd *TestVCD) testCreateOrgVdcNetworkRouted(check *C, ipSubnet string, sub
 	connectedGw, err := vcd.vdc.FindEdgeGatewayNameByNetwork(networkName)
 	check.Assert(err, IsNil)
 	check.Assert(connectedGw, Equals, edgeGWName)
+
+	networkId := network.OrgVDCNetwork.ID
+	updatedNetworkName := networkName + "-updated"
+	updatedNetworkDescription := "Updated by govcd tests"
+	updatedStartAddress := ipSubnet + ".5"
+	updatedEndAddress := ipSubnet + ".30"
+	network.OrgVDCNetwork.Name = updatedNetworkName
+	network.OrgVDCNetwork.Description = updatedNetworkDescription
+	network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].StartAddress = updatedStartAddress
+	network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].EndAddress = updatedEndAddress
+
+	err = network.Update()
+	check.Assert(err, IsNil)
+	AddToCleanupList(updatedNetworkName,
+		"network",
+		vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name,
+		"Test_CreateOrgVdcNetworkRouted")
+
+	network, err = vcd.vdc.GetOrgVdcNetworkById(networkId, true)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, updatedNetworkName)
+	check.Assert(network.OrgVDCNetwork.Description, Equals, updatedNetworkDescription)
+	check.Assert(network.OrgVDCNetwork.Configuration, NotNil)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes, NotNil)
+	check.Assert(len(network.OrgVDCNetwork.Configuration.IPScopes.IPScope), Not(Equals), 0)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].Gateway, Equals, gateway)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].StartAddress, Equals, updatedStartAddress)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].EndAddress, Equals, updatedEndAddress)
 
 	task, err := network.Delete()
 	check.Assert(err, IsNil)
@@ -180,8 +216,8 @@ func (vcd *TestVCD) Test_GetNetworkList(check *C) {
 	check.Assert(found, Equals, true)
 }
 
-// Tests the creation of an isolated Org VDC network
-func (vcd *TestVCD) Test_CreateOrgVdcNetworkIso(check *C) {
+// Tests the creation and update of an isolated Org VDC network
+func (vcd *TestVCD) Test_CreateUpdateOrgVdcNetworkIso(check *C) {
 	fmt.Printf("Running: %s\n", check.TestName())
 	networkName := TestCreateOrgVdcNetworkIso
 
@@ -190,52 +226,113 @@ func (vcd *TestVCD) Test_CreateOrgVdcNetworkIso(check *C) {
 		check.Skip(fmt.Sprintf("Error deleting network : %s", err))
 	}
 
-	var networkConfig = types.OrgVDCNetwork{
-		Name: networkName,
-		// Description: "Created by govcd tests",
-		Configuration: &types.NetworkConfiguration{
-			FenceMode: types.FenceModeIsolated,
-			/*One of:
-				bridged (connected directly to the ParentNetwork),
-			  isolated (not connected to any other network),
-			  natRouted (connected to the ParentNetwork via a NAT service)
-			  https://code.vmware.com/apis/287/vcloud#/doc/doc/types/OrgVdcNetworkType.html
-			*/
-			IPScopes: &types.IPScopes{
-				IPScope: []*types.IPScope{&types.IPScope{
-					IsInherited: false,
-					Gateway:     "192.168.2.1",
-					Netmask:     "255.255.255.0",
-					DNS1:        "192.168.2.1",
-					DNS2:        "",
-					DNSSuffix:   "",
-					IPRanges: &types.IPRanges{
-						IPRange: []*types.IPRange{
-							&types.IPRange{
-								StartAddress: "192.168.2.2",
-								EndAddress:   "192.168.2.100"},
+	var (
+		gateway             = "192.168.2.1"
+		updatedNetworkName  = TestCreateOrgVdcNetworkIso + "updated"
+		startAddress        = "192.168.2.2"
+		updatedStartAddress = "192.168.2.5"
+		endAddress          = "192.168.2.100"
+		updatedEndAddress   = "192.168.2.50"
+		netmask             = "255.255.255.0"
+		dns1                = gateway
+		dns2                = "8.8.8.8"
+		dnsSuffix           = "vcloud.org"
+		description         = "Created by govcd tests"
+		updatedDescription  = "Updated by govcd tests"
+		networkConfig       = types.OrgVDCNetwork{
+			Xmlns:       types.XMLNamespaceVCloud,
+			Name:        networkName,
+			Description: description,
+			Configuration: &types.NetworkConfiguration{
+				FenceMode: types.FenceModeIsolated,
+				/*One of:
+					bridged (connected directly to the ParentNetwork),
+				  isolated (not connected to any other network),
+				  natRouted (connected to the ParentNetwork via a NAT service)
+				  https://code.vmware.com/apis/287/vcloud#/doc/doc/types/OrgVdcNetworkType.html
+				*/
+				IPScopes: &types.IPScopes{
+					IPScope: []*types.IPScope{&types.IPScope{
+						IsInherited: false,
+						Gateway:     gateway,
+						Netmask:     netmask,
+						DNS1:        dns1,
+						DNS2:        dns2,
+						DNSSuffix:   dnsSuffix,
+						IPRanges: &types.IPRanges{
+							IPRange: []*types.IPRange{
+								&types.IPRange{
+									StartAddress: startAddress,
+									EndAddress:   endAddress,
+								},
+							},
 						},
 					},
+					},
 				},
-				}},
-			BackwardCompatibilityMode: true,
-		},
-		IsShared: false,
-	}
+				BackwardCompatibilityMode: true,
+			},
+			IsShared: false,
+		}
+	)
+
 	LogNetwork(networkConfig)
 	err = vcd.vdc.CreateOrgVDCNetworkWait(&networkConfig)
 	if err != nil {
 		fmt.Printf("error creating Network <%s>: %s\n", networkName, err)
 	}
 	check.Assert(err, IsNil)
-	AddToCleanupList(TestCreateOrgVdcNetworkIso,
+	AddToCleanupList(networkName,
 		"network",
 		vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name,
 		"Test_CreateOrgVdcNetworkIso")
+
+	network, err := vcd.vdc.GetOrgVdcNetworkByName(networkName, true)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Description, Equals, description)
+	check.Assert(network.OrgVDCNetwork.Configuration.FenceMode, Equals, types.FenceModeIsolated)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes, NotNil)
+	check.Assert(len(network.OrgVDCNetwork.Configuration.IPScopes.IPScope), Not(Equals), 0)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].Gateway, Equals, gateway)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].Netmask, Equals, netmask)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].DNS1, Equals, dns1)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].DNS2, Equals, dns2)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].DNSSuffix, Equals, dnsSuffix)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges, NotNil)
+	check.Assert(len(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange), Not(Equals), 0)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].StartAddress, Equals, startAddress)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].EndAddress, Equals, endAddress)
+
+	networkId := network.OrgVDCNetwork.ID
+	network.OrgVDCNetwork.Description = updatedDescription
+	network.OrgVDCNetwork.Name = updatedNetworkName
+	network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].StartAddress = updatedStartAddress
+	network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].EndAddress = updatedEndAddress
+	err = network.Update()
+	check.Assert(err, IsNil)
+	AddToCleanupList(updatedNetworkName,
+		"network",
+		vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name,
+		"Test_CreateOrgVdcNetworkIso")
+
+	network, err = vcd.vdc.GetOrgVdcNetworkById(networkId, true)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, updatedNetworkName)
+	check.Assert(network.OrgVDCNetwork.Description, Equals, updatedDescription)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes, NotNil)
+	check.Assert(len(network.OrgVDCNetwork.Configuration.IPScopes.IPScope), Not(Equals), 0)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].Netmask, Equals, netmask)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].DNS2, Equals, dns2)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].DNSSuffix, Equals, dnsSuffix)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges, NotNil)
+	check.Assert(len(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange), Not(Equals), 0)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].StartAddress, Equals, updatedStartAddress)
+	check.Assert(network.OrgVDCNetwork.Configuration.IPScopes.IPScope[0].IPRanges.IPRange[0].EndAddress, Equals, updatedEndAddress)
+
 }
 
-// Tests the creation of a Org VDC network connected to an external network
-func (vcd *TestVCD) Test_CreateOrgVdcNetworkDirect(check *C) {
+// Tests the creation and update of a Org VDC network connected to an external network
+func (vcd *TestVCD) Test_CreateUpdateOrgVdcNetworkDirect(check *C) {
 	fmt.Printf("Running: %s\n", check.TestName())
 	networkName := TestCreateOrgVdcNetworkDirect
 
@@ -256,8 +353,11 @@ func (vcd *TestVCD) Test_CreateOrgVdcNetworkDirect(check *C) {
 		return
 	}
 	// Note that there is no IPScope for this type of network
+	description := "Created by govcd test"
 	var networkConfig = types.OrgVDCNetwork{
-		Name: networkName,
+		Xmlns:       types.XMLNamespaceVCloud,
+		Name:        networkName,
+		Description: description,
 		Configuration: &types.NetworkConfiguration{
 			FenceMode: types.FenceModeBridged,
 			ParentNetwork: &types.Reference{
@@ -281,7 +381,7 @@ func (vcd *TestVCD) Test_CreateOrgVdcNetworkDirect(check *C) {
 	}
 	check.Assert(task.Task.HREF, Not(Equals), "")
 
-	AddToCleanupList(TestCreateOrgVdcNetworkDirect,
+	AddToCleanupList(networkName,
 		"network", vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name,
 		"Test_CreateOrgVdcNetworkDirect")
 
@@ -292,22 +392,142 @@ func (vcd *TestVCD) Test_CreateOrgVdcNetworkDirect(check *C) {
 	}
 	check.Assert(err, IsNil)
 
+	// Retrieving the network
+	newNetwork, err := vcd.vdc.GetOrgVdcNetworkByName(networkName, true)
+	check.Assert(err, IsNil)
+	check.Assert(newNetwork, NotNil)
+	check.Assert(newNetwork.OrgVDCNetwork.Name, Equals, networkName)
+	check.Assert(newNetwork.OrgVDCNetwork.Description, Equals, description)
+	check.Assert(newNetwork.OrgVDCNetwork.Configuration, NotNil)
+	check.Assert(newNetwork.OrgVDCNetwork.Configuration.ParentNetwork, NotNil)
+	check.Assert(newNetwork.OrgVDCNetwork.Configuration.ParentNetwork.ID, Equals, externalNetwork.ExternalNetwork.ID)
+	check.Assert(newNetwork.OrgVDCNetwork.Configuration.ParentNetwork.Name, Equals, externalNetwork.ExternalNetwork.Name)
+
+	// Updating
+	updatedNetworkName := networkName + "-updated"
+	updatedDescription := "Updated by govcd tests"
+	newNetwork.OrgVDCNetwork.Description = updatedDescription
+	newNetwork.OrgVDCNetwork.Name = updatedNetworkName
+	err = newNetwork.Update()
+	check.Assert(err, IsNil)
+
+	AddToCleanupList(updatedNetworkName,
+		"network", vcd.org.Org.Name+"|"+vcd.vdc.Vdc.Name,
+		"Test_CreateOrgVdcNetworkDirect")
+
+	// Check the values of the updated entities. The new values should be available
+	// immediately.
+	check.Assert(newNetwork.OrgVDCNetwork.Name, Equals, updatedNetworkName)
+	check.Assert(newNetwork.OrgVDCNetwork.Description, Equals, updatedDescription)
+
+	// Gets a new copy of the network and check the values.
+	newNetwork, err = vcd.vdc.GetOrgVdcNetworkByName(updatedNetworkName, true)
+	check.Assert(err, IsNil)
+	check.Assert(newNetwork, NotNil)
+	check.Assert(newNetwork.OrgVDCNetwork.Name, Equals, updatedNetworkName)
+	check.Assert(newNetwork.OrgVDCNetwork.Description, Equals, updatedDescription)
+
+	// restoring original name
+	newNetwork.OrgVDCNetwork.Name = networkName
+	err = newNetwork.Update()
+	check.Assert(err, IsNil)
+	check.Assert(newNetwork.OrgVDCNetwork.Name, Equals, networkName)
+
 	// Testing RemoveOrgVdcNetworkIfExists:
 	// (1) Make sure the network exists
-	newNetwork, err := vcd.vdc.GetOrgVdcNetworkByName(TestCreateOrgVdcNetworkDirect, true)
+	newNetwork, err = vcd.vdc.GetOrgVdcNetworkByName(networkName, true)
 	check.Assert(err, IsNil)
 	check.Assert(newNetwork, NotNil)
 
 	// (2) Removing the network. It should return nil, as a successful deletion
-	err = RemoveOrgVdcNetworkIfExists(*vcd.vdc, TestCreateOrgVdcNetworkDirect)
+	err = RemoveOrgVdcNetworkIfExists(*vcd.vdc, networkName)
 	check.Assert(err, IsNil)
 
 	// (3) Look for the network again. It should be deleted
-	_, err = vcd.vdc.GetOrgVdcNetworkByName(TestCreateOrgVdcNetworkDirect, true)
+	_, err = vcd.vdc.GetOrgVdcNetworkByName(networkName, true)
 	check.Assert(err, NotNil)
 	check.Assert(IsNotFound(err), Equals, true)
 
 	// (4) Attempting a second conditional deletion. It should also return nil, as the network was not found
-	err = RemoveOrgVdcNetworkIfExists(*vcd.vdc, TestCreateOrgVdcNetworkDirect)
+	err = RemoveOrgVdcNetworkIfExists(*vcd.vdc, networkName)
 	check.Assert(err, IsNil)
+}
+
+func (vcd *TestVCD) Test_NetworkUpdateRename(check *C) {
+	if vcd.config.VCD.Network.Net1 == "" {
+		check.Skip("Skipping test because no network was given")
+	}
+
+	fmt.Printf("Running: %s\n", check.TestName())
+
+	network, err := vcd.vdc.GetOrgVdcNetworkByName(vcd.config.VCD.Network.Net1, false)
+
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, vcd.config.VCD.Network.Net1)
+	var saveNetwork = types.OrgVDCNetwork{
+		HREF:        network.OrgVDCNetwork.HREF,
+		Type:        network.OrgVDCNetwork.Type,
+		ID:          network.OrgVDCNetwork.ID,
+		Name:        network.OrgVDCNetwork.Name,
+		Status:      network.OrgVDCNetwork.Status,
+		Description: network.OrgVDCNetwork.Description,
+	}
+
+	// Change name and description
+	updatedNetworkName := "UpdatedNetwork"
+	updatedDescription := "Updated description"
+	network.OrgVDCNetwork.Description = updatedDescription
+	network.OrgVDCNetwork.Name = updatedNetworkName
+	err = network.Update()
+	check.Assert(err, IsNil)
+
+	// Retrieve the network again
+	network, err = vcd.vdc.GetOrgVdcNetworkById(saveNetwork.ID, false)
+	check.Assert(err, IsNil)
+
+	check.Assert(network.OrgVDCNetwork.Name, Equals, updatedNetworkName)
+	check.Assert(network.OrgVDCNetwork.Description, Equals, updatedDescription)
+	check.Assert(network.OrgVDCNetwork.HREF, Equals, saveNetwork.HREF)
+	check.Assert(network.OrgVDCNetwork.Type, Equals, saveNetwork.Type)
+	check.Assert(network.OrgVDCNetwork.ID, Equals, saveNetwork.ID)
+	check.Assert(network.OrgVDCNetwork.EdgeGateway, DeepEquals, saveNetwork.EdgeGateway)
+	check.Assert(network.OrgVDCNetwork.Status, Equals, saveNetwork.Status)
+
+	// Change back name and description to their original values
+	network.OrgVDCNetwork.Description = saveNetwork.Description
+	network.OrgVDCNetwork.Name = saveNetwork.Name
+
+	err = network.Update()
+	check.Assert(err, IsNil)
+	network, err = vcd.vdc.GetOrgVdcNetworkById(saveNetwork.ID, false)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Description, Equals, saveNetwork.Description)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, saveNetwork.Name)
+
+	// An update without any changes should run without producing side effects
+	err = network.Update()
+	check.Assert(err, IsNil)
+	network, err = vcd.vdc.GetOrgVdcNetworkById(saveNetwork.ID, false)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Description, Equals, saveNetwork.Description)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, saveNetwork.Name)
+
+	// We run some of the above operations using Rename instead of Update
+	err = network.Rename(updatedNetworkName)
+	check.Assert(err, IsNil)
+	network, err = vcd.vdc.GetOrgVdcNetworkById(saveNetwork.ID, false)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, updatedNetworkName)
+
+	// Trying to rename with the same name should give an error
+	err = network.Rename(updatedNetworkName)
+	check.Assert(err, NotNil)
+
+	// Setting the original name again
+	err = network.Rename(saveNetwork.Name)
+	check.Assert(err, IsNil)
+	network, err = vcd.vdc.GetOrgVdcNetworkById(saveNetwork.ID, false)
+	check.Assert(err, IsNil)
+	check.Assert(network.OrgVDCNetwork.Name, Equals, saveNetwork.Name)
+
 }